+++ docs/README.md
@@ -1,4 +1,4 @@
-# Bubbles User Guide
+# bubbles.Bubbles User Guide
 
 // Update the title above to match the actual product name
