+++ src/main/java/bubbles/BubblesException.java
@@ -1,3 +1,5 @@
+package bubbles;
+
 public class BubblesException extends Exception{
     public BubblesException(String message){
         super(message);
