+++ src/main/java/bubbles/Command.java
@@ -1,3 +1,5 @@
+package bubbles;
+
 public abstract class Command {
     private String command;
     public Command(String command){
