+++ src/main/java/bubbles/Event.java
@@ -1,3 +1,5 @@
+package bubbles;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
