+++ src/main/java/bubbles/EventCommand.java
@@ -1,3 +1,5 @@
+package bubbles;
+
 import java.io.IOException;
 
 public class EventCommand extends Command{
