+++ src/main/java/bubbles/Parser.java
@@ -1,3 +1,5 @@
+package bubbles;
+
 import java.io.IOException;
 
 public class Parser {
