+++ src/main/java/bubbles/Task.java
@@ -1,3 +1,5 @@
+package bubbles;
+
 public class Task {
     protected String description;
     protected boolean isDone;
