+++ text-ui-test/EXPECTED.TXT
@@ -1,4 +1,4 @@
-Hello! I'm Bubbles 🫧 
+Hello! I'm bubbles.Bubbles 🫧
 What can I do for you?
 --------------------------------------------------------
 --------------------------------------------------------
